Handle filtered out file messages from krazy2.
When krazy2all is run, it silently filters out the files that will not
be checked (because it can not be accessed, or because its file type is
not supported). However, krazy2 prints a message explaining why it
filtered out a file. ProgressParser has been updated to take into
account those messages.