Add support for non-ASCII file names.
Before, the parser classes expected the data as a QString. However, when
the process output, a QByteArray, was implicitly converted to a QString,
the fromAscii method was used. krazy2 writes its XML ouput as UTF-8, so
the non-ASCII file names were not parsed properly by ResultParser, as
the internal QXMLStreamReader got a bad encoded QString. Thus, the
QStrings returned by the QXMLStreamReader were bad encoded too.

Now, the parsers expect QByteArrays. The internal QXMLStreamReader in
ResultParser receives QByteArrays too, so it takes care of using the
proper encoding based on the XML header, and thus the QStrings returned
by the QXMLStreamReader are properly encoded.

The ProgressParser does not need a proper encoding, as the progress data
is all output in ASCII; only error messages may contain non-ASCII
characters, and those messages are discarded. However, it has been
changed for consistency with the ResultParser.