Use only krazy2 instead of krazy2 and krazy2all.
Before, krazy2all was ran in the directory to be analyzed. krazy2all is
just a wrapper that gets all the files in the directory (and,
recursively, its subdirectories), silently filters out the files that
will not be checked, and then runs krazy2 with the names of the actual
files to be checked.

Now, instead of running krazy2all, the recursive list of files in the
directory to be analized is got and passed to krazy2. krazy2 itself
automatically filters out the files that will not be checked before
performing the analysis.