Add AnalysisParameters class.
Before, the AnalysisJob executed krazy2 with the default checkers on all
the files of a single directory. Moreover, it got the list of available
checkers every time it was run just to calculate the number of checkers
that were going to be run.

Now, the AnalysisParameters class stores the checkers to run and the
paths to be analyzed. Several files and directories can be analyzed in
the same run using only the specified checkers (even extra checkers).
Moreover, the available checkers are kept in the AnalysisParameters, so
they need to be initialized only once. If the AnalysisParameters passed
to the AnalysisJob was not initialized yet, the AnalysisJob initializes
it before performing the actual analysis.