Add selection of checkers to run.
Before, only the default Krazy2 checkers were run. Now, it can be
selected the checkers to run.

If the dialog to select the checkers is shown before the checkers were
initialized, the job to initialze the checkers will be started and the
dialog will be updated once the checkers are initialized.