Fix running every checker with a duplicated name.
Before, if a checker X for a file type A was selected to run, every
checker X for any file type was run. Now, only the selected checkers are
run. To do this, due to the krazy2 command options behavior, krazy2 has
to be executed several times in AnalysisJob, each time running the
checkers for only one file type.