Fix wrong markup.
The KUIT markup has to be used with i18nc in order to be converted to
plain text or rich text, depending on the context. Here no
internationalization is needed, so just wrapping the text in <p>
elements to make the text to be treated like rich text is enough.