Copy the checkers from the analysis parameters to the results.
Before, the checkers were added to the results by parsing the krazy2
output. However, krazy2 does not specify in the results whether a
checker is extra or not. So, now, the checkers are copied from the
analysis parameters to the results before the parsing, as the checkers
coming from the parameters are fully initialized.