Add support in AnalysisJob for several AnalysisParameters.
Before, only a single set of AnalysisParameters could be used in an
AnalysisJob. Now, several AnalysisParameters can be added to be executed
in the same batch.

The execution of each AnalysisParameters will be like executing several
AnalysisJobs, each one with a single AnalysisParameters. However, the
AnalysisResults will contain the results for all the analysis, and the
progress report will take into account all the checkers to be run from
all the AnalysisParameters.