Add a consistent sorting for the issues.
Before, the sorting of the issues in the view depended not just on the
column to sort by, but also on the previous columns that was sorted by
(as QSortFilterProxyModel uses a stable sort, and thus keeps the sorting
if two items have the same value). Now, the sorting of each column
depends also on the other columns, but now in a controlled way, so the
sorting will be always the same, no matter what other columns were the
issues sorted by previously.