Add a context menu to analyze again the selected issues or files.
When analyzing again issues, the checker of each issue to analyze again
is run on the file of its issue. When analyzing again files, all the
checkers selected by the user in the general analysis are run on the
selected files. In both cases, the results of the new analyzis are
merged into the previous results.