Use custom tool tips for the results view.
Before, the default tool tips were used in the results view. The tool
tip text provided by the model are the explanations of the issues.
However, the explanations may contain links, but they could not be
clicked or copied as the tool tip was hidden when the cursor moved into
it. Now, the tool tip is shown using ActiveToolTip class from
kdevplatform, which is not hidden when the cursor enters into it.

Unfortunately, no automated tests are provided, as it is too cumbersome
to test this :(